moved queryrepository to repositories.py
