+++ backend/infrastructures/apps/load/adapters/__init__.py
@@ -1,9 +1,8 @@
-from .queries import DjangoDataQueryRepository
-from .repositories import DjangoDataDomainRepository
+from .repositories import DjangoDataDomainRepository, DjangoDataQueryRepository
 from .units_of_work import DjangoDataUnitOfWork
 
 __all__ = [
-    "DjangoDataUnitOfWork",
     "DjangoDataDomainRepository",
     "DjangoDataQueryRepository",
+    "DjangoDataUnitOfWork",
 ]
