+++ backend/infrastructures/config/settings/injections_configurations.py
@@ -1,6 +1,6 @@
 import inject
 
-from ...apps.load.adapters import queries as query_load_repositories
+from ...apps.load.adapters import repositories as query_load_repositories
 from ...apps.load.adapters import units_of_work as load_unit_of_work
